fix(skill): correct evals.json mapping and reduce to 2 evals
- Fix eval-002 to correctly map AdditiveModel to eval_2_target.py
- Rename eval_3_target.py -> eval_2_target.py (AdditiveModel)
- Delete unused eval_3 and eval_4 target files (PersistenceModel, HistoryModel)
- Update ground_truth description with more detail
