+++ skills/earth2studio-create-prognostic/evals/targets/eval_2_target.py
@@ -14,10 +14,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Persistence prognostic model that predicts future by returning current state.
+"""Additive prognostic model that adds a constant at each time step.
 
-A persistence forecast assumes conditions remain unchanged. Useful as a
-simple baseline for weather prediction.
+Demonstrates normalization buffers (center and scale) loaded from package.
 """
 
 from collections import OrderedDict
@@ -33,30 +32,48 @@ from earth2studio.models.px.utils import PrognosticMixin
 from earth2studio.utils import handshake_coords, handshake_dim
 from earth2studio.utils.type import CoordSystem
 
-VARIABLES = ["t2m", "msl", "z500", "t850", "u500", "v500"]
+VARIABLES = ["t2m", "u10m", "v10m", "msl"]
 
 
-class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
-    """Persistence prognostic model.
+class AdditiveModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
+    """Additive prognostic model.
 
-    A baseline model that predicts the future by assuming the current state
-    persists unchanged. Useful as a reference for forecast skill evaluation.
+    A simple model that adds a small constant (0.01) to each variable at
+    each time step. Demonstrates normalization with center/scale buffers.
 
     Parameters
     ----------
     core_model : torch.nn.Module, optional
         Core model (not used, for interface compatibility).
+    center : torch.Tensor, optional
+        Normalization center values.
+    scale : torch.Tensor, optional
+        Normalization scale values.
 
     Note
     ----
-    Persistence forecasts are a common baseline in weather prediction.
+    This model demonstrates the use of normalization buffers in E2S wrappers.
     """
 
-    def __init__(self, core_model: torch.nn.Module | None = None) -> None:
+    def __init__(
+        self,
+        core_model: torch.nn.Module | None = None,
+        center: torch.Tensor | None = None,
+        scale: torch.Tensor | None = None,
+    ) -> None:
         super().__init__()
         self.model = core_model
+        if center is not None:
+            self.register_buffer("center", center)
+        else:
+            self.register_buffer("center", torch.zeros(len(VARIABLES)))
+        if scale is not None:
+            self.register_buffer("scale", scale)
+        else:
+            self.register_buffer("scale", torch.ones(len(VARIABLES)))
         self.register_buffer("device_buffer", torch.empty(0))
-        self._time_step = np.timedelta64(24, "h")
+        self._time_step = np.timedelta64(6, "h")
+        self._increment = 0.01
 
     def input_coords(self) -> CoordSystem:
         """Input coordinate system of the prognostic model.
@@ -72,8 +89,8 @@ class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
                 "time": np.empty(0),
                 "lead_time": np.array([np.timedelta64(0, "h")]),
                 "variable": np.array(VARIABLES),
-                "lat": np.linspace(90, -90, 721, endpoint=True),
-                "lon": np.linspace(0, 360, 1440, endpoint=False),
+                "lat": np.linspace(90, -90, 361, endpoint=True),
+                "lon": np.linspace(0, 360, 720, endpoint=False),
             }
         )
 
@@ -115,7 +132,7 @@ class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
         Returns
         -------
         Package
-            Model package (empty for persistence model).
+            Model package.
         """
         return Package(".")
 
@@ -126,14 +143,46 @@ class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
         Parameters
         ----------
         package : Package
-            Model package (not used for persistence model).
+            Model package with model.pt, center.npy, scale.npy.
 
         Returns
         -------
         PrognosticModel
             Loaded model instance.
         """
-        return cls(core_model=None)
+        # For testing without files, use defaults
+        try:
+            model_path = package.resolve("model.pt")
+            core_model = torch.load(model_path, map_location="cpu", weights_only=False)
+            core_model.eval()
+        except Exception:
+            core_model = None
+
+        try:
+            center_path = package.resolve("center.npy")
+            center = torch.from_numpy(np.load(center_path))
+        except Exception:
+            center = None
+
+        try:
+            scale_path = package.resolve("scale.npy")
+            scale = torch.from_numpy(np.load(scale_path))
+        except Exception:
+            scale = None
+
+        return cls(core_model, center=center, scale=scale)
+
+    def _normalize(self, x: torch.Tensor) -> torch.Tensor:
+        """Normalize input tensor."""
+        center = self.center.view(1, 1, 1, -1, 1, 1)
+        scale = self.scale.view(1, 1, 1, -1, 1, 1)
+        return (x - center) / scale
+
+    def _denormalize(self, x: torch.Tensor) -> torch.Tensor:
+        """Denormalize output tensor."""
+        center = self.center.view(1, 1, 1, -1, 1, 1)
+        scale = self.scale.view(1, 1, 1, -1, 1, 1)
+        return x * scale + center
 
     @batch_func()
     def __call__(
@@ -153,7 +202,7 @@ class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
         Returns
         -------
         tuple[torch.Tensor, CoordSystem]
-            Output tensor (persistence) and coordinates one time step ahead.
+            Output tensor with added constant and coordinates one time step ahead.
         """
         target_input_coords = self.input_coords()
         handshake_coords(coords, target_input_coords, "variable")
@@ -162,8 +211,11 @@ class PersistenceModel(torch.nn.Module, AutoModelMixin, PrognosticMixin):
         device = self.device_buffer.device
         x = x.to(device)
 
+        # Add small constant
+        output = x + self._increment
+
         out_coords = self.output_coords(coords)
-        return x, out_coords
+        return output, out_coords
 
     @batch_func()
     def _default_generator(
