+++ skills/earth2studio-create-prognostic/evals/targets/test/eval_2_test_target.py
@@ -14,7 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Tests for PersistenceModel prognostic model."""
+"""Tests for AdditiveModel prognostic model."""
 
 from collections import OrderedDict
 from collections.abc import Iterable
@@ -24,18 +24,30 @@ import pytest
 import torch
 
 from earth2studio.data import Random, fetch_data
-from earth2studio.models.px import PersistenceModel
+from earth2studio.models.auto import Package
+from earth2studio.models.px import AdditiveModel
 from earth2studio.utils import handshake_dim
 
 
-class PhooPersistenceModel(torch.nn.Module):
-    """Dummy persistence model for testing."""
+class PhooAdditiveModel(torch.nn.Module):
+    """Dummy additive model for testing."""
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        return x
+        return x + 0.01
 
 
-class TestPersistenceModelMock:
+@pytest.fixture(scope="class")
+def test_package(tmp_path_factory):
+    """Create dummy package with model and normalization files."""
+    tmp_path = tmp_path_factory.mktemp("model_data")
+    model = PhooAdditiveModel()
+    torch.save(model, tmp_path / "model.pt")
+    np.save(tmp_path / "center.npy", np.zeros(4))
+    np.save(tmp_path / "scale.npy", np.ones(4))
+    return Package(str(tmp_path))
+
+
+class TestAdditiveModelMock:
     """Mock tests using dummy model weights."""
 
     @pytest.mark.parametrize(
@@ -56,9 +68,9 @@ class TestPersistenceModelMock:
             ),
         ],
     )
-    def test_persistence_call(self, time, device):
-        """Test single forward pass."""
-        model = PersistenceModel.load_model(PersistenceModel.load_default_package())
+    def test_additive_call(self, test_package, time, device):
+        """Test single forward pass adds constant."""
+        model = AdditiveModel.load_model(test_package)
         model = model.to(device)
 
         dc = model.input_coords()
@@ -73,8 +85,10 @@ class TestPersistenceModelMock:
         assert isinstance(out_coords, OrderedDict)
         handshake_dim(out_coords, "variable", 3)
 
-        # 24h time step
-        expected_lead = coords["lead_time"] + np.timedelta64(24, "h")
+        # Check that output is input + 0.01
+        torch.testing.assert_close(out, x + 0.01)
+
+        expected_lead = coords["lead_time"] + np.timedelta64(6, "h")
         np.testing.assert_array_equal(out_coords["lead_time"], expected_lead)
 
     @pytest.mark.parametrize("ensemble", [1, 2])
@@ -90,9 +104,9 @@ class TestPersistenceModelMock:
             ),
         ],
     )
-    def test_persistence_iter(self, ensemble, device):
-        """Test iterator produces correct sequence."""
-        model = PersistenceModel.load_model(PersistenceModel.load_default_package())
+    def test_additive_iter(self, test_package, ensemble, device):
+        """Test iterator produces correct sequence with accumulating additions."""
+        model = AdditiveModel.load_model(test_package)
         model = model.to(device)
 
         time = np.array([np.datetime64("2024-01-01T00:00")])
@@ -108,18 +122,18 @@ class TestPersistenceModelMock:
         iterator = model.create_iterator(x, coords)
         assert isinstance(iterator, Iterable)
 
-        # Check initial condition (step 0)
         step0_x, step0_coords = next(iterator)
         assert step0_x.shape[0] == ensemble
         np.testing.assert_array_equal(
             step0_coords["lead_time"], np.array([np.timedelta64(0, "h")])
         )
 
-        # Check subsequent steps with 24h time step
         for i, (step_x, step_coords) in enumerate(iterator):
             assert step_x.shape[0] == ensemble
-            expected_lead = np.array([np.timedelta64((i + 1) * 24, "h")])
+            expected_lead = np.array([np.timedelta64((i + 1) * 6, "h")])
             np.testing.assert_array_equal(step_coords["lead_time"], expected_lead)
+            # Additive model accumulates: x + (i+1) * 0.01
+            torch.testing.assert_close(step_x, x + (i + 1) * 0.01)
             if i >= 4:
                 break
 
@@ -138,9 +152,9 @@ class TestPersistenceModelMock:
             ),
         ],
     )
-    def test_persistence_exceptions(self, coords):
+    def test_additive_exceptions(self, test_package, coords):
         """Test model raises on invalid coordinates."""
-        model = PersistenceModel.load_model(PersistenceModel.load_default_package())
+        model = AdditiveModel.load_model(test_package)
         x = torch.randn(
             1,
             1,
@@ -152,9 +166,17 @@ class TestPersistenceModelMock:
         with pytest.raises((KeyError, ValueError)):
             model(x, coords)
 
-    def test_persistence_grid_size(self):
-        """Test grid size is 721x1440 (0.25 degree)."""
-        model = PersistenceModel.load_model(PersistenceModel.load_default_package())
+    def test_additive_normalization_buffers(self, test_package):
+        """Test normalization buffers are loaded from package."""
+        model = AdditiveModel.load_model(test_package)
+        assert hasattr(model, "center")
+        assert hasattr(model, "scale")
+        assert model.center.shape == (4,)
+        assert model.scale.shape == (4,)
+
+    def test_additive_grid_size(self):
+        """Test grid size is 361x720 (0.5 degree)."""
+        model = AdditiveModel.load_model(AdditiveModel.load_default_package())
         coords = model.input_coords()
-        assert len(coords["lat"]) == 721
-        assert len(coords["lon"]) == 1440
+        assert len(coords["lat"]) == 361
+        assert len(coords["lon"]) == 720
